+++ src/createRelease.ts
@@ -2,15 +2,16 @@ import * as core from '@actions/core';
 import { getOctokit } from '@actions/github';
 import { octobase, vars } from './constants';
 import { createTagName } from './createTagName';
-import { getReleaseDescription } from './getReleaseDescription';
+import { getReleaseNotes } from './getReleaseNotes';
 
 export async function createRelease(
   version: string
 ): Promise<void> {
-  const notesSource = core.getInput('notes-source', { required: false });
   const tag = createTagName(version);
-  const body = notesSource === 'changelog'
-  ? await getReleaseDescription(version) : undefined;
+  const {
+    body,
+    generate_release_notes,
+  } = await getReleaseNotes(version);
   const octokit = getOctokit(vars.token);
 
   try {
@@ -19,10 +20,10 @@ export async function createRelease(
       tag_name: tag,
       name: tag,
       body,
-      generate_release_notes: notesSource === 'auto',
+      generate_release_notes,
     });
 
-    core.info(`Created release ${releaseData.name} 🏷️`);
+    core.info(`🏷️ Created release tag ${releaseData.name}`);
   } catch (error) {
     console.error(error);
     core.setFailed(