+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.2.0 (12-03-2023)
+
+**Features:**
+- specify fallback for release notes when no changelog file founded or when no description in this file for the new version with `notes-fallback` input: `fallbackText` | `auto`
+- specify custom text that used for release notes fallback with `fallback-text` input
+
+**Fixes:**
+- fix #1: unable to read files from repo in which that action is running
+
 ## 0.1.0 (06-03-2023)
 
 > `generate-release` now available to use as `beta`